requirements.txt
Django>=4.0.3,<4.1
psycopg2>=2.9.3,<3.0
uWSGI>=2.0.20,<2.1
djangorestframework>=3.13.1,<3.14
django-simple-history>=3.0.0,<3.1
